image-service/requirements.txt
Flask==3.0.0
Pillow==10.1.0
rawpy==0.18.1
imageio==2.31.5
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
